Redis, Lazysizes, XSS, and other security & performance improvements

+++ Client/src/components/helpers.js
@@ -1,28 +1,33 @@
-export const fetchData = async (target) => {
+export const fetchData = async (target, post = false) => {
   const res = await fetch(`http://localhost:5000/${target}`,{
+    method: post ? 'POST' : 'GET',
     credentials: 'include'
   });
-  return await res.json();
+  return res.status === 200 ? await res.json() : [];
+};
+
+const fetchPost = async (target) => {
+  const res = await fetch(`http://localhost:5000/${target}`,{
+    method: 'POST',
+    credentials: 'include'
+  });
+  return res.status === 200 ? true : false;
 };
 
 export const logout = async (setAuthenticated) => {
-  await fetch("http://localhost:5000/auth/logout", {
-    method: "POST",
-    credentials: 'include'
-  });
-  setAuthenticated(false);
-  localStorage.clear();
+  const res = await fetchPost("auth/logout");
+  if (res) {
+    setAuthenticated(false);
+    localStorage.clear();
+  }
 };
 
 export const handleSubscription = async (id, action) => {
-    const response = await fetchData(`news/${action}/${id}`);
+    const response = await fetchData(`news/${action}/${id}`, true);
     localStorage.setItem("sources", JSON.stringify(response));
 }
 
 
 export const checkAuthenticated = async () => {
-  const res = await fetch('http://localhost:5000/auth/checkAuth',{
-    credentials: 'include'
-  });
-  return res.status === 200 ? true : false;
+  return await fetchPost("auth/checkAuth");
 }
